voice_cloner/voice-cloning-server/src/utils/model_loader.py
"""
模型加载和推理模块 - 简化版

由于实际的 SV2TTS 模型需要较大的文件，这里提供模拟实现
实际部署时需要：
1. 下载预训练模型
2. 替换为真实的模型加载代码
"""

import numpy as np
from loguru import logger
from typing import Optional
import torch


class MockEncoder:
    """
    模拟的 Encoder (用于测试)
    实际部署时需要加载真实的 SV2TTS Encoder 模型
    """
    
    def __init__(self, model_path: Optional[str] = None):
        self.model_path = model_path
        self.is_loaded = False
        logger.info("初始化 Mock Encoder")
    
    def load_model(self, model_path: str):
        """加载模型"""
        logger.info(f"加载 Encoder 模型: {model_path}")
        # 实际部署时需要加载真实模型
        # self.model = torch.load(model_path)
        self.is_loaded = True
        logger.info("✓ Encoder 模型加载成功 (Mock)")
    
    def embed_utterance(self, audio: np.ndarray) -> np.ndarray:
        """
        提取声纹特征
        
        Args:
            audio: 音频数据 (16kHz, float32)
            
        Returns:
            声纹向量 (256维)
        """
        if not self.is_loaded:
            raise RuntimeError("模型未加载")
        
        logger.info(f"提取声纹特征，音频长度: {len(audio)}")
        
        # 模拟实现：生成一个基于音频统计特征的向量
        # 实际部署时需要使用真实的模型推理
        
        # 简单的特征提取（模拟）
        mean = np.mean(audio)
        std = np.std(audio)
        energy = np.sum(audio ** 2)
        
        # 生成256维向量（模拟）
        # 使用音频的统计特征作为种子，保证相同音频生成相同向量
        seed = int((abs(mean) + abs(std) + abs(energy)) * 1000000) % 2**32
        np.random.seed(seed)
        embedding = np.random.randn(256).astype(np.float32)
        
        # 归一化
        embedding = embedding / (np.linalg.norm(embedding) + 1e-6)
        
        logger.info(f"✓ 声纹提取完成，维度: {embedding.shape}")
        return embedding


class MockSynthesizer:
    """
    模拟的 Synthesizer (用于测试)
    实际部署时需要加载真实的 SV2TTS Synthesizer 模型
    """
    
    def __init__(self, model_path: Optional[str] = None):
        self.model_path = model_path
        self.is_loaded = False
        logger.info("初始化 Mock Synthesizer")
    
    def load_model(self, model_path: str):
        """加载模型"""
        logger.info(f"加载 Synthesizer 模型: {model_path}")
        # 实际部署时需要加载真实模型
        self.is_loaded = True
        logger.info("✓ Synthesizer 模型加载成功 (Mock)")
    
    def synthesize_spectrograms(self, texts: list, embeddings: list) -> list:
        """
        合成梅尔频谱图
        
        Args:
            texts: 文本列表
            embeddings: 声纹向量列表
            
        Returns:
            频谱图列表
        """
        if not self.is_loaded:
            raise RuntimeError("模型未加载")
        
        logger.info(f"合成频谱图，文本数量: {len(texts)}")
        
        spectrograms = []
        for text, embedding in zip(texts, embeddings):
            # 模拟生成频谱图
            # 实际部署时需要使用真实的模型推理
            
            # 根据文本长度生成对应长度的频谱图（模拟）
            text_length = len(text)
            time_steps = max(50, text_length * 10)  # 模拟时间步
            mel_bins = 80  # 梅尔频谱的频率维度
            
            # 生成随机频谱图（模拟）
            spec = np.random.randn(mel_bins, time_steps).astype(np.float32)
            spectrograms.append(spec)
        
        logger.info(f"✓ 频谱图合成完成")
        return spectrograms


class MockVocoder:
    """
    模拟的 Vocoder (用于测试)
    实际部署时需要加载真实的 HiFi-GAN 或 WaveGlow 模型
    """
    
    def __init__(self, model_path: Optional[str] = None):
        self.model_path = model_path
        self.is_loaded = False
        logger.info("初始化 Mock Vocoder")
    
    def load_model(self, model_path: str):
        """加载模型"""
        logger.info(f"加载 Vocoder 模型: {model_path}")
        # 实际部署时需要加载真实模型
        self.is_loaded = True
        logger.info("✓ Vocoder 模型加载成功 (Mock)")
    
    def infer_waveform(self, spectrogram: np.ndarray) -> np.ndarray:
        """
        从频谱图生成音频波形
        
        Args:
            spectrogram: 梅尔频谱图
            
        Returns:
            音频波形 (16kHz)
        """
        if not self.is_loaded:
            raise RuntimeError("模型未加载")
        
        logger.info(f"生成音频波形，频谱图形状: {spectrogram.shape}")
        
        # 模拟生成音频
        # 实际部署时需要使用真实的模型推理
        
        # 根据频谱图时间步生成对应长度的音频（模拟）
        time_steps = spectrogram.shape[1]
        sample_rate = 16000
        hop_length = 200  # 常见的hop length
        
        audio_length = time_steps * hop_length
        
        # 生成简单的正弦波（模拟语音）
        t = np.linspace(0, audio_length / sample_rate, audio_length)
        frequencies = [200, 400, 600, 800]  # 模拟多个频率成分
        audio = np.zeros(audio_length)
        for freq in frequencies:
            audio += np.sin(2 * np.pi * freq * t) * 0.2
        
        # 添加一些包络变化（模拟语音的强度变化）
        envelope = np.abs(np.sin(np.linspace(0, np.pi * 10, audio_length)))
        audio = audio * envelope
        
        audio = audio.astype(np.float32)
        
        logger.info(f"✓ 音频生成完成，长度: {len(audio)/sample_rate:.2f}秒")
        return audio


# 全局模型实例
_encoder = None
_synthesizer = None
_vocoder = None


def get_encoder() -> MockEncoder:
    """获取 Encoder 实例"""
    global _encoder
    if _encoder is None:
        _encoder = MockEncoder()
    return _encoder


def get_synthesizer() -> MockSynthesizer:
    """获取 Synthesizer 实例"""
    global _synthesizer
    if _synthesizer is None:
        _synthesizer = MockSynthesizer()
    return _synthesizer


def get_vocoder() -> MockVocoder:
    """获取 Vocoder 实例"""
    global _vocoder
    if _vocoder is None:
        _vocoder = MockVocoder()
    return _vocoder


def load_all_models(encoder_path: str, synthesizer_path: str, vocoder_path: str):
    """加载所有模型"""
    logger.info("开始加载所有模型...")
    
    encoder = get_encoder()
    synthesizer = get_synthesizer()
    vocoder = get_vocoder()
    
    encoder.load_model(encoder_path)
    synthesizer.load_model(synthesizer_path)
    vocoder.load_model(vocoder_path)
    
    logger.info("✓ 所有模型加载完成")
