voice_cloner/voice-cloning-server/requirements.txt
# FastAPI 和 Web 服务器
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据处理和验证
pydantic==2.5.0
python-dotenv==1.0.0

# 音频处理
librosa==0.10.1
soundfile==0.12.1
resampy==0.4.2
scipy==1.11.4
numpy==1.24.3

# 深度学习（可选，Mock模式不需要）
# torch==2.1.0
# torchaudio==2.1.0

# 日志
loguru==0.7.2

# 中文处理（用于TTS）
jieba==0.42.1
pypinyin==0.50.0

# 其他工具
requests==2.31.0
tqdm==4.66.1
